Add files via upload
Included the images that were used for the project as I did not see them in the repository, and I believe that when it is ran externally it mightn't work. I couldn't test this as VSCode is being a pain as per usual, but if this is incorrect we can just remove them before the next part of the assignment. Also would be better to put these into a folder but alas. My thing doesn't work right now.